schema/pyproject.toml
[project]
name = "lsap-schema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "observerw", email = "wozluohd@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "lsprotocol>=2025.0.0",
    "pydantic>=2.12.5",
    "python-liquid>=2.1.0",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
